# swagger.yaml
openapi: 3.0.0
info:
  title: Cirkle API
  version: 1.0.0
  description: |
    ## Cirkle — Social Events Platform API

    Cirkle is an AI-native social discovery platform for people to find their tribe and attend events together.

    ### Authentication
    All endpoints except `GET /cities`, `GET /lifestyle-tags`, and `POST /auth/login` require a Bearer token.

    Include the token in every protected request:
    Authorization: Bearer <your_jwt_token>

    ### Price Format
    All prices are in **paise** (₹1 = 100 paise). Example: ₹500 = `50000`.

    ### Phone Format
    Indian mobile numbers only. Format: `+91XXXXXXXXXX` where the 10-digit number starts with 6, 7, 8, or 9.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /admin/admins:
    post:
      tags: [Admin — Admins]
      summary: Create an admin account
      description: |
        Creates a new admin (business_development or administrative). This is how administrative admins create BD team accounts.

        Requires the manage_admins capability — a business_development admin gets 403.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
      responses:
        '201':
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error (email format, password length, missing displayName, invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability (business_development role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: You do not have permission to manage admin accounts
        '409':
          description: An admin with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Admin — Admins]
      summary: List all admins
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of admin accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  admins:
                    type: array
                    items:
                      $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/admins/{id}:
    get:
      tags: [Admin — Admins]
      summary: Get a single admin account
      description: Same shape as the list rows. password_hash is never returned.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Admin — Admins]
      summary: Update an admin account
      description: |
        Partial update — only fields present in the body are changed. Commonly used to deactivate an admin (`isActive: false`), which invalidates their existing token immediately (authenticateAdmin re-checks is_active on every request, not just at token issuance).

        **Lockout protection.** Two rules make admin management unbrickable. Both return 409 with a machine-readable `error` so the client can say why:

        - `cannot_deactivate_self` — nobody may deactivate their own account, unconditionally, however many other admins exist. The failure mode is losing the session you are in.
        - `last_administrative_admin` — the population of ACTIVE administrative admins may never reach zero, whether by deactivating the last one or demoting them to business_development.

        Self-demotion is deliberately NOT blanket-blocked: it is governed by the second rule alone, so an administrative admin may hand over and step down to BD as long as somebody else remains administrative.

        The population count is taken under a row lock in the same transaction as the update, so two admins demoting each other simultaneously cannot both pass and leave zero.

        **Password reset.** Send `password` to reset it — alone, or alongside other fields. A body containing only `password` is valid and no longer returns "No fields to update". The value is bcrypt-hashed, never returned and never logged; the audit row records only that a reset occurred, not the value. A reset changes no counts, so the lockout rules never block it — resetting your own password is always allowed.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
                isActive:
                  type: boolean
                password:
                  type: string
                  minLength: 8
                  format: password
                  description: >-
                    New password. Hashed with bcrypt before storage, never
                    returned in any response and never written to logs. May be
                    the only field in the body.
      responses:
        '200':
          description: Admin updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error, or no fields provided to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Blocked by lockout protection. Branch on `error`:

            - `cannot_deactivate_self` — you tried to deactivate your own account.
            - `last_administrative_admin` — the change would leave zero active administrative admins.
            - `concurrent_admin_change` — a competing admin change collided; retry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [cannot_deactivate_self, last_administrative_admin, concurrent_admin_change]
                  message:
                    type: string
                    description: Human-readable. Branch on `error`, not on this.
              examples:
                self:
                  value:
                    error: cannot_deactivate_self
                    message: You cannot deactivate your own account.
                last:
                  value:
                    error: last_administrative_admin
                    message: There must be at least one active administrative admin.
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/users:
    get:
      tags: [Admin — Users]
      summary: Search and list users
      description: |
        Search matches phone, email and first/last name — case-insensitive partial.
        **PII masking by role.** End-user `phone` and `email` are role-gated. An `administrative` admin holds `view_pii` and receives the real values; a `business_development` admin receives MASKED ones — phone as `+91******210` (country code plus last three digits), email as `j****@gmail.com` (first character plus the domain, which is kept for context).

        Masking is applied server-side, so the real value never enters the response body for a BD admin and cannot be recovered from the network tab. There is no reveal endpoint or query flag.

        Search matches the REAL stored values, so a BD admin can still find a user by phone or email; the results come back masked.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
          description: Matched case-insensitively, partial, against phone, email, first name, and last name
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminUserSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/{id}:
    get:
      tags: [Admin — Users]
      summary: Full user detail — the support-lookup screen
      description: |
        Complete profile (presigned photo URLs, lifestyle tags, bio, tagline), phone, email, derived age, plus full order and ticket history.

        **PII masking by role.** End-user `phone` and `email` are role-gated: `administrative` admins hold `view_pii` and see real values, `business_development` admins receive masked ones (`+91******210`, `j****@gmail.com`). Masking is server-side, so the real value never enters the response body.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full user detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/AdminUserDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login. Valid for 7 days.
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
      required: false
      description: Max rows to return. Default 50, clamped to 100 even if a larger value is requested.
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message here
    Admin:
      type: object
      description: password_hash is never included in any response.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: admin@cirkle.live
        displayName:
          type: string
          example: Founder Admin
        role:
          type: string
          enum: [administrative, business_development]
          description: Only capability difference is manage_admins — administrative has it, business_development doesn't. Otherwise identical access.
        isActive:
          type: boolean
          description: Kill switch. A deactivated admin's existing token stops working immediately, not just at expiry. Absent from the login response (implied true — login already checked it).
        createdAt:
          type: string
          format: date-time
...
    PaginationEnvelope:
      type: object
      description: Shared shape for every list endpoint. `data` is added per-endpoint via allOf.
      properties:
        total:
          type: integer
          description: Total matching rows, ignoring limit/offset — for "showing X-Y of Z" and page controls
          example: 3847
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 100
...
    AdminUserSummary:
      type: object
      description: |
        PII note: phone and email are returned unmasked to every admin for
        now. A future view_pii capability will mask these for
        business_development admins, done server-side. All raw exposure is
        funneled through two functions in src/routes/adminUsers.js so that
        change touches one place, not every endpoint.
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        gender:
          type: string
          nullable: true
        cityId:
          type: string
          nullable: true
        phone:
          type: string
        email:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        ticketCount:
          type: integer
        orderCount:
          type: integer
    AdminUserDetail:
      type: object
      description: Same PII note as AdminUserSummary.
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        gender:
          type: string
          nullable: true
        cityId:
          type: string
          nullable: true
        phone:
          type: string
        email:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        tagline:
          type: string
          nullable: true
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        lifestyleTags:
          type: array
          items:
            $ref: '#/components/schemas/LifestyleTag'
        createdAt:
          type: string
          format: date-time
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              status:
                type: string
              totalPaise:
                type: integer
              createdAt:
                type: string
                format: date-time
              eventName:
                type: string
        tickets:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              bookingRef:
                type: string
              checkedIn:
                type: boolean
              createdAt:
                type: string
                format: date-time
              eventName:
                type: string
              eventStartsAt:
                type: string
                format: date-time
...
    LifestyleTag:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: be1c46e0-e44b-4776-89fd-e15b97b95fc7
        label:
          type: string
          example: Clubbing
        category:
          type: string
          example: Going out
...
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          description: Presigned GET URL, ~1 hour TTL. The bucket is private — never cache this long-term, it expires. Generated fresh on every read.
          example: https://cirkle-user-photos.s3.ap-south-1.amazonaws.com/profiles/user-id/photo-0.jpg?X-Amz-Signature=...
        position:
          type: integer
          minimum: 0
          maximum: 3
          description: Position 0 is the Main photo
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: No token provided
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...